requirements.txt
python-binance
pyyaml
pandas